ParkingLotStatus/pyproject.toml
[project]
name = "parkinglotstatus"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.3.2",
    "opencv-python>=4.11.0.86",
    "pillow>=11.3.0",
    "ultralytics>=8.3.179",
]
